requirements.txt
alchemy-mock==0.4.3
alembic==1.7.5
astroid==2.12.10
attrs==21.4.0
autopep8==2.0.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
dill==0.3.5.1
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Breadcrumbs==0.5.1
Flask-Login==0.5.0
Flask-Menu==0.7.2
Flask-Migrate==3.1.0
Flask-MySQL==1.5.2
Flask-SQLAlchemy==2.5.1
genson==1.2.2
greenlet==1.1.2
importlib-metadata==4.9.0
importlib-resources==5.4.0
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
Js2Py==0.71
jsonschema==4.3.3
lazy-object-proxy==1.7.1
lxml==4.8.0
Mako==1.1.6
MarkupSafe==2.0.1
mccabe==0.7.0
mypy==0.982
mypy-extensions==0.4.3
packaging==21.3
paramiko==2.8.1
Pillow==9.0.1
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
py-mini-racer==0.6.0
pycodestyle==2.9.1
pycparser==2.21
pyjsparser==2.7.1
pylint==2.15.3
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==3.0.9
pyrsistent==0.18.0
pytest==7.1.3
pytest-flask-sqlalchemy==1.1.0
pytest-mock==3.10.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==1.4.41
sqlalchemy-datatables==2.0.1
sshtunnel==0.4.0
svgutils==0.3.4
tomli==2.0.1
tomlkit==0.11.5
typing_extensions==4.4.0
tzdata==2021.5
tzlocal==4.1
Werkzeug==2.0.2
wrapt==1.14.1
zipp==3.6.0
